requirements.txt
#conda install pytorch==1.11.0 cudatoolkit=11.3 -c pytorch
g2p-en
inflect
librosa
matplotlib
numba
numpy
pypinyin
pyworld
PyYAML
scikit-learn
scipy
soundfile
tensorboard
tgt
tqdm
unidecode
pandas
openpyxl